+++ auth_service/app/alembic/versions/2025_07_18_1521-a574160513b6_add_role_table.py
@@ -1,0 +1,38 @@
+"""add_role_table
+
+Revision ID: a574160513b6
+Revises: 
+Create Date: 2025-07-18 15:21:14.531399
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'a574160513b6'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        'role',
+        sa.Column('id', sa.Integer(), nullable=False),
+        sa.Column('name', sa.Enum('user', 'manager', 'admin', native_enum=False), nullable=False),
+        sa.Column('description', sa.String(), nullable=True),
+        sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('role')
+    # ### end Alembic commands ###
