+++ auth_service/app/alembic/versions/2025_07_18_1526-699d8a9477d7_add_permission_table.py
@@ -1,0 +1,41 @@
+"""add_permission_table
+
+Revision ID: 699d8a9477d7
+Revises: c4881e859291
+Create Date: 2025-07-18 15:26:20.451206
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '699d8a9477d7'
+down_revision: Union[str, Sequence[str], None] = 'c4881e859291'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        'permission',
+        sa.Column('id', sa.Integer(), nullable=False),
+        sa.Column('code', sa.String(), nullable=False),
+        sa.Column('description', sa.String(), nullable=False),
+        sa.PrimaryKeyConstraint('id'),
+        sa.UniqueConstraint('code')
+    )
+    op.create_unique_constraint(None, 'user', ['id'])
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint(None, 'user', type_='unique')
+    op.drop_table('permission')
+    # ### end Alembic commands ###
