+++ auth_service/app/alembic/versions/2025_07_18_1528-8d311faa9b9c_add_roles_permissions_table.py
@@ -1,0 +1,39 @@
+"""add_roles_permissions_table
+
+Revision ID: 8d311faa9b9c
+Revises: 699d8a9477d7
+Create Date: 2025-07-18 15:28:01.141054
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '8d311faa9b9c'
+down_revision: Union[str, Sequence[str], None] = '699d8a9477d7'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('roles_permissions',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('role_id', sa.Integer(), nullable=False),
+    sa.Column('permission_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['permission_id'], ['permission.id'], ),
+    sa.ForeignKeyConstraint(['role_id'], ['role.id'], ),
+    sa.PrimaryKeyConstraint('id', 'role_id', 'permission_id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('roles_permissions')
+    # ### end Alembic commands ###
